requirements.txt
libvirt-python
flask
ansible
ansible-lint
lxml
